2.5.STABLE9-alt1
- upstream merged:
 + 2005-02-20 10:47 (Minor) Relax header parsing slightly again to work around broken web servers
 + 2005-02-20 19:11 (Cosmetic) GCC4 warnings
 + 2005-02-21 01:38 (Cosmetic) Doesn't work specifying the AR variable to configure
 + 2005-02-21 02:58 (Minor) Peer related memory leaks on "squid -k reconfigure"
 + 2005-02-21 03:38 (Cosmetic) Display FTP URLs in decoded format to allow for sane display of national characters etc
 + 2005-02-21 17:02 (Minor) fqdn lookups with spaces may confuse redirectors
 + 2005-02-23 00:11 (Medium) Should not automatically retry request on 403 and other server errors
- upstream updated and merged:
 + 2005-02-20 11:03 (Cosmetic) Cross-platform format fixes
- updated FAQ to v 1.245 2005/02/24 23:29:59

+++ squid/src/net_db.c
@@ -1,6 +1,6 @@
 
 /*
- * $Id: net_db.c,v 1.158.2.4 2003/01/02 23:21:57 wessels Exp $
+ * $Id: net_db.c,v 1.158.2.5 2005/02/13 21:23:19 serassio Exp $
  *
  * DEBUG: section 38    Network Measurement Database
  * AUTHOR: Duane Wessels
@@ -621,7 +621,7 @@ netdbExchangeHandleReply(void *data, char *buf, ssize_t size)
     }
     debug(38, 3) ("netdbExchangeHandleReply: used %d entries, (x %d bytes) == %d bytes total\n",
 	nused, rec_sz, nused * rec_sz);
-    debug(38, 3) ("netdbExchangeHandleReply: seen %d, used %d\n", ex->seen, ex->used);
+    debug(38, 3) ("netdbExchangeHandleReply: seen %ld, used %ld\n", (long int) ex->seen, (long int) ex->used);
     if (EBIT_TEST(ex->e->flags, ENTRY_ABORTED)) {
 	debug(38, 3) ("netdbExchangeHandleReply: ENTRY_ABORTED\n");
 	netdbExchangeDone(ex);
